Feat: 캔버스 이미지 드로잉, 저장 기능 커스텀 훅으로 분리
